requirements.txt
langchain==0.0.346
openai==0.28.1
python-dotenv==1.0.0
pytest==7.4.0